plot api: move plot/update sub-routine to subroutines.js
- this move will allow us to test (with jasmine spies)
  that the proper subroutines are called.
